requirements.txt
pandas
matplotlib
pyodbc
python-dotenv
sqlalchemy